# docs/api.yaml
...
openapi: 3.1.1
servers:
  - url: http://127.0.0.1:4002/v1
info:
  title: A-Novel Story Schematics API.
  description: |
    The A-Novel story schematics API, for internal use by A-Novel services.
  contact:
    name: A-Novel
    email: support@agoradesecrivains.com
  license:
    name: GNU AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  version: 0.0.0
security:
  - bearerAuth: []

# ======================================================================================================================
# Tags
# ======================================================================================================================
...
paths:
...
  /beats-sheet:
    put:
      tags:
        - beats-sheet
      security:
        - bearerAuth:
            - "beats-sheet:create"
      summary: Create a new beats sheet.
      description: |
        Create a new beats sheet for a logline, following a story plan.
      operationId: createBeatsSheet
      requestBody:
        $ref: "#/components/requestBodies/CreateBeatsSheetForm"
      responses:
        "200":
          description: The beats sheet was created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeatsSheet"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404":
          description: The logline or story plan does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "422":
          description: The beats sheet does not match the story plan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnprocessableEntityError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
    get:
      tags:
        - beats-sheet
      security:
        - bearerAuth:
            - "beats-sheet:read"
      summary: Get a beats sheet.
      description: |
        Get a beats sheet.
      operationId: getBeatsSheet
      parameters:
        - $ref: "#/components/parameters/BeatsSheetID"
      responses:
        "200":
          description: The beats sheet was retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeatsSheet"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404":
          description: The beats sheet does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /beats-sheets:
    get:
      tags:
        - beats-sheet
      security:
        - bearerAuth:
            - "beats-sheets:read"
      summary: Get all beats sheets.
      description: |
        Get all beats sheets for the current user.
      operationId: getBeatsSheets
      parameters:
        - $ref: "#/components/parameters/LoglineID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: The beats sheets were retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BeatsSheetPreview"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /beats-sheet/generate:
    post:
      tags:
        - beats-sheet
      security:
        - bearerAuth:
            - "beats-sheet:generate"
      summary: Generate a new beats sheet.
      description: |
        Generate a new beats sheet for a logline, following a story plan.
      operationId: generateBeatsSheet
      requestBody:
        $ref: "#/components/requestBodies/GenerateBeatsSheetForm"
      responses:
        "200":
          description: The beats sheet was generated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeatsSheetIdea"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404":
          description: The logline or story plan does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /beats-sheet/regenerate:
    post:
      tags:
        - beats-sheet
      security:
        - bearerAuth:
            - "beats-sheet:regenerate"
      summary: Regenerate beats in a beats sheet.
      description: |
        Regenerate the content of specific beats in a beats sheet.
      operationId: regenerateBeats
      requestBody:
        $ref: "#/components/requestBodies/RegenerateBeatsForm"
      responses:
        "200":
          description: The beats were regenerated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Beats"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404":
          description: The beats sheet does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /beats-sheet/expand:
    post:
      tags:
        - beats-sheet
      security:
        - bearerAuth:
            - "beat:expand"
      summary: Expand a beat in a beats sheet.
      description: |
        Add more details to a specific beat in a beats sheet.
      operationId: expandBeat
      requestBody:
        $ref: "#/components/requestBodies/ExpandBeatForm"
      responses:
        "200":
          description: The beat was expanded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Beat"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404":
          description: The beats sheet does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "422":
          description: The beat does not exist in the beats sheet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnprocessableEntityError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /logline:
    put:
      tags:
        - logline
      security:
        - bearerAuth:
            - "logline:create"
      summary: Create a new logline.
      description: |
        Create a new logline for a user.
      operationId: createLogline
      requestBody:
        $ref: "#/components/requestBodies/CreateLoglineForm"
      responses:
        "200":
          description: The logline was created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Logline"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
    get:
      tags:
        - logline
      security:
        - bearerAuth:
            - "logline:read"
      summary: Get a logline.
      description: |
        Get a logline.
      operationId: getLogline
      parameters:
        - in: query
          name: id
          required: false
          schema:
            $ref: "#/components/schemas/LoglineID"
        - in: query
          name: slug
          required: false
          schema:
            $ref: "#/components/schemas/Slug"
      responses:
        "200":
          description: The logline was retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Logline"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404":
          description: The logline does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /loglines:
    get:
      tags:
        - logline
      security:
        - bearerAuth:
            - "loglines:read"
      summary: Get all loglines.
      description: |
        Get all loglines for the current user.
      operationId: getLoglines
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: The loglines were retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoglinePreview"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /loglines/generate:
    post:
      tags:
        - logline
      security:
        - bearerAuth:
            - "loglines:generate"
      summary: Generate new loglines.
      description: |
        Generate new loglines for a user.
      operationId: generateLoglines
      requestBody:
        $ref: "#/components/requestBodies/GenerateLoglinesForm"
      responses:
        "200":
          description: The loglines were generated successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoglineIdea"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /logline/expand:
    post:
      tags:
        - logline
      security:
        - bearerAuth:
            - "logline:expand"
      summary: Expand a logline idea.
      description: |
        Add more details to a logline idea.
      operationId: expandLogline
      requestBody:
        $ref: "#/components/requestBodies/ExpandLoglineForm"
      responses:
        "200":
          description: The logline was expanded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoglineIdea"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
...
  /story-plans:
    get:
      tags:
        - story-plan
      security:
        - bearerAuth:
            - "story-plans:read"
      summary: Get all story plans.
      description: |
        Get all story plans.
      operationId: getStoryPlans
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: The story plans were retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoryPlanPreview"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"

# ======================================================================================================================
# Components
# ======================================================================================================================
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
  # ======================================================= SCHEMAS ====================================================
  schemas:
...
    BeatsSheet:
      type: object
      required:
        - id
        - loglineID
        - storyPlanID
        - content
        - lang
        - createdAt
      description: A beats sheet is a detailed outline of a story, breaking it down into its individual beats.
      properties:
        id:
          $ref: "#/components/schemas/BeatsSheetID"
        loglineID:
          $ref: "#/components/schemas/LoglineID"
        storyPlanID:
          $ref: "#/components/schemas/StoryPlanID"
        content:
          type: array
          maxItems: 128
          items:
            $ref: "#/components/schemas/Beat"
        lang:
          $ref: "#/components/schemas/Lang"
          description: The language of the beats sheet.
          example: en
        createdAt:
          type: string
          format: date-time
          description: The date and time at which the beats sheet was created.
          example: 2022-01-01T00:00:00Z
...
    UnauthorizedError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message.
          example: The provided access token is not valid.
...
  requestBodies:
    CreateBeatsSheetForm:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CreateBeatsSheetForm"
...
  parameters:
    LoglineID:
      name: loglineID
      in: query
      required: true
      description: The unique identifier of the logline.
      schema:
        $ref: "#/components/schemas/LoglineID"
...